Bug correction and missing functionality (stayed behind...teehee)
Corrected the showing of messages when there are no finished or no
unfinished tasks by adding "loadingFinished" and "loadingUnfinished" to
the TaskValue provider.

Added delete dialog to Finished Tasks button.

Minor alterations.

+++ app/components/dialog/controller/TaskVisualizationDialogCtrl.js
@@ -9,11 +9,29 @@
         vm.task = task;
         
         //Triggered by the create button
-        function deleteTask() {
-        	ToastrService.processing("Deleting", "Please wait while the task is deleted...");
-            TaskService.deleteTask(task)
-                .then(function () {closeDialog(); ToastrService.success("Task succesfully deleted.");})
-                .catch(function (error) {ToastrService.error("Error", "There was a problem in the deletion. Please refresh the page before trying again.");console.log('TaskVisualizationDialogCtrl error (deleteTask): ', error);});
+        function deleteTask(event) {
+
+            DialogService.openDeleteConfirmationDialog(event)
+                .then(function (answer) {
+
+                    //Answer can be yes or no. If yes, then proceed with the delete operation, otherwise, do nothing.
+                    if (answer === 'yes') {
+                        closeDialog();
+                        ToastrService.processing("Deleting", "Please wait while the task is deleted...");
+                        TaskService.deleteTask(task)
+                            .then(function () {
+                                ToastrService.success("Task succesfully deleted.");
+                            })
+                            .catch(function (error) {
+                                ToastrService.error("Error", "There was a problem in the deletion. Please refresh the page before trying again.");
+                                console.log('TaskListCtrl error (deleteTask): ', error);
+                            });
+                    }
+
+                }, function () {
+                    //Delete cancelled
+                });
+
         }
         
         function closeDialog() {
